feat: countdown ticker runs at all hover times, not just under 50s
Previously _shouldCountdownTick(const Duration(minutes: 1)) gated the
hover countdown ticker, so above 1 minute the display was frozen to the
snapshot value.  Now _hasCountdownData starts the ticker whenever either
intervalRemains or weeklyRemains is non-null — the ticker decrements the
anchor-based countdown regardless of magnitude.  The render object's
dirty-check absorbs frames where the formatted string doesn't change
(e.g. within the same minute), so the ticker stays cheap.

Also reorders onEnter so _syncCountdownTicker (anchor) runs before
_pushCurrentFrame, preventing a single frozen frame on hover entry.
_pushCurrentFrame gains a countdown-ticker branch so hover entry
immediately renders the anchored value.

+++ lib/src/models/session_runtime_state.dart
@@ -72,7 +72,7 @@ class SessionRuntimeState {
   int accumulatedToolTokens;
   String thinkingMode;
   String? reasoningEffort;
-  int? cacheHitPct;
+  double? cacheHitPct;
 
   bool isGeneratingTldr;
